Refactor articles - Move to separate module - Get rid of some duplication
