+++ lib/blawg_postgres_db/helpers.ex
@@ -1,0 +1,9 @@
+defmodule BlawgPostgresDb.Helpers do
+  def humanize_errors(changeset) do
+    Ecto.Changeset.traverse_errors(changeset, fn {msg, opts} ->
+      Enum.reduce(opts, msg, fn {key, value}, acc ->
+        String.replace(acc, "%{#{key}}", to_string(value))
+      end)
+    end)
+  end
+end
